+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.2 (Upcoming)
  - Made it possible to use arrow keys to select command shown in command palette (`CTRL+SHIFT+P`)
+ - Made it possible to request input with options
 
 ## 1.1.1
  - Added options to close tabs in context menu when right-clicking on a tab